# Serialisation: text and binary