--- a/CHANGELOG.md
+++ b/CHANGELOG.md
@@ -8,7 +8,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 <!-- ---------------------
       v1.6.0
      --------------------- -->
-## v1.6.0 - Unreleased
+## v1.6.0 - 2-12-2024
 
 ### Added