--- a/README.md
+++ b/README.md
@@ -248,7 +248,8 @@ The `DTensor` `B` will be overwritten with the solution.
 
 ### 1.8. Saving and loading tensors
 
-Tensor data can be stored in simple text files which have the following structure
+Tensor data can be stored in simple text files or binary files. 
+The text-based format has the following structure
 
 ```text
 number_of_rows
@@ -259,13 +260,15 @@ data (one entry per line)
 
 To save a tensor in a file, simply call `DTensor::saveToFile(filename)`.
 
-To load a tensor from a file, the static function `DTensor<T>::parseFromTextFile(filename)` can be used. For example:
+If the file extension is `.bt` (binary tensor), the data will be stored in binary format.
+
+To load a tensor from a file, the static function `DTensor<T>::parseFromFile(filename)` can be used. For example:
 
 ```c++
-auto z = DTensor<double>::parseFromTextFile("path/to/my.dtensor")
+auto z = DTensor<double>::parseFromFile("path/to/my.dtensor")
 ```
 
-If necessary, you can provide a second argument to `parseFromTextFile` to specify the order in which the data are stored (the `StorageMode`). 
+If necessary, you can provide a second argument to `parseFromFile` to specify the order in which the data are stored (the `StorageMode`). 
 
 ## 2. Cholesky factorisation and system solution