--- a/README.md
+++ b/README.md
@@ -261,6 +261,9 @@ data (one entry per line)
 To save a tensor in a file, simply call `DTensor::saveToFile(filename)`.
 
 If the file extension is `.bt` (binary tensor), the data will be stored in binary format.
+The structure of the binary encoding is similar to that of the text encoding:
+the first three `uint64_t`-sized positions correspond to the number of rows, columns
+and matrices, followed by the elements of the tensor.
 
 To load a tensor from a file, the static function `DTensor<T>::parseFromFile(filename)` can be used. For example:
 
@@ -268,7 +271,9 @@ To load a tensor from a file, the static function `DTensor<T>::parseFromFile(fil
 auto z = DTensor<double>::parseFromFile("path/to/my.dtensor")
 ```
 
-If necessary, you can provide a second argument to `parseFromFile` to specify the order in which the data are stored (the `StorageMode`). 
+If necessary, you can provide a second argument to `parseFromFile` to specify the order in which the data are stored (the `StorageMode`).
+
+Soon we will release a Python API for reading and serialising (numpy) arrays to `.bt` files.  
 
 ## 2. Cholesky factorisation and system solution