--- a/main.cu
+++ b/main.cu
@@ -6,9 +6,15 @@
 
 
 int main() {
-    auto z = DTensor<size_t>::parseFromTextFile("../test/data/my.dtensor",
-                                                StorageMode::rowMajor);
-    std::cout << z;
-    z.saveToFile("hohoho.dtensor");
+    /* Write to binary file */
+    auto r = DTensor<double>::createRandomTensor(3, 6, 4, -1, 1);
+    std::string fName = "tensor.bt"; // binary tensor file extension: .bt
+    r.saveToFile(fName);
+
+    /* Parse binary file */
+    auto recov = DTensor<double>::parseFromFile(fName);
+    auto err = r - recov;
+    std::cout << "max error : " << err.maxAbs();
+
     return 0;
 }